chore: add miss rate data in scripts

+++ task1/scripts/GeneralRun.sh
@@ -16,7 +16,7 @@ SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)"
 PROJECT_DIR="$(cd -- "$SCRIPT_DIR/.." && pwd)"
 CSV="${PROJECT_DIR}/results/${OPERATION}_results.csv"
 
-MATRIX_SIZES=(256 512 752 1024 1256 1504 1752 2048)
+MATRIX_SIZES=(1024 1256 1504 1752 2048 2560 3072 3584 4096)
 
 # =========================================================
 # Get sudo permission once
@@ -32,7 +32,7 @@ sudo -v || exit 1
 # =========================================================
 
 mkdir -p "${PROJECT_DIR}/results"
-echo "matrix_size,instructions,l1d_misses,speedup" > "$CSV"
+echo "matrix_size,instructions,l1d_misses,l1d_loads,speedup" > "$CSV"
 
 
 # =========================================================
@@ -57,7 +57,8 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
         sudo perf stat \
             -e cpu_core/cycles/ \
             -e cpu_core/instructions/ \
-            -e cpu_core/l1d_miss.load/ \
+            -e cpu_core/L1-dcache-load-misses/ \
+            -e cpu_core/L1-dcache-loads/ \
             "$PROJECT_DIR/bin/conv" "$OPERATION" "$MATRIXSIZE" "$MATRIXSIZE" 7 50 \
             > "$OUTPUT" 2>&1
 
@@ -82,7 +83,11 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
         # Extract L1-D misses
         # -------------------------------------------------
 
-        L1D_MISSES=$(grep 'cpu_core/l1d_miss.load' "$OUTPUT" \
+        L1D_MISSES=$(grep 'cpu_core/L1-dcache-load-misses' "$OUTPUT" \
+            | sed 's/,//g' \
+            | awk '{print $1}')
+
+        L1D_LOADS=$(grep 'cpu_core/L1-dcache-loads' "$OUTPUT" \
             | sed 's/,//g' \
             | awk '{print $1}')
 
@@ -99,12 +104,12 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
         # Save to CSV
         # -------------------------------------------------
 
-        if [[ "$INSTRUCTIONS" =~ ^[0-9]+$ && "$L1D_MISSES" =~ ^[0-9]+$ ]]; then
-            echo "$MATRIXSIZE,$INSTRUCTIONS,$L1D_MISSES,$SPEEDUP" >> "$CSV"
+        if [[ "$INSTRUCTIONS" =~ ^[0-9]+$ && "$L1D_MISSES" =~ ^[0-9]+$ && "$L1D_LOADS" =~ ^[0-9]+$ ]]; then
+            echo "$MATRIXSIZE,$INSTRUCTIONS,$L1D_MISSES,$L1D_LOADS,$SPEEDUP" >> "$CSV"
             rm -f "$OUTPUT"
             break
         else
-            echo "Invalid instructions or L1-D misses value; rerunning matrix size $MATRIXSIZE"
+            echo "Invalid instructions, L1-D misses, or L1-D loads value; rerunning matrix size $MATRIXSIZE"
             rm -f "$OUTPUT"
         fi
     done
@@ -113,6 +118,7 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
     echo "Captured:"
     echo "  Instructions : $INSTRUCTIONS"
     echo "  L1-D misses  : $L1D_MISSES"
+    echo "  L1-D loads   : $L1D_LOADS"
     echo "  Speedup      : $SPEEDUP"
 
 done
