+++ task1/scripts/SIMDRun.sh
@@ -5,7 +5,7 @@ PROJECT_DIR="$(cd -- "$SCRIPT_DIR/.." && pwd)"
 SOURCE="$PROJECT_DIR/src/conv_simd.cpp"
 CSV="$PROJECT_DIR/results/simd_results.csv"
 
-MATRIX_SIZES=(256 512 752 1024 1256 1504 1752 2048)
+MATRIX_SIZES=(1024 1256 1504 1752 2048 2560 3072 3584 4096)
 
 # SIMD register widths to sweep, mapped to the is_simd_256 boolean
 # 128 -> false (SSE), 256 -> true (AVX2)
@@ -24,7 +24,7 @@ sudo -v || exit 1
 # Create CSV
 # =========================================================
 
-echo "matrix_size,simd_register_width,instructions,l1d_misses,speedup" > "$CSV"
+echo "matrix_size,simd_register_width,instructions,l1d_misses,l1d_loads,speedup" > "$CSV"
 
 
 # =========================================================
@@ -79,7 +79,8 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
             sudo perf stat \
                 -e cpu_core/cycles/ \
                 -e cpu_core/instructions/ \
-                -e cpu_core/l1d_miss.load/ \
+                -e cpu_core/L1-dcache-load-misses/ \
+                -e cpu_core/L1-dcache-loads/ \
                 "$PROJECT_DIR/bin/conv" simd "$MATRIXSIZE" "$MATRIXSIZE" 7 50 \
                 > "$OUTPUT" 2>&1
 
@@ -104,7 +105,11 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
             # Extract L1-D misses
             # -------------------------------------------------
 
-            L1D_MISSES=$(grep 'cpu_core/l1d_miss.load' "$OUTPUT" \
+            L1D_MISSES=$(grep 'cpu_core/L1-dcache-load-misses' "$OUTPUT" \
+                | sed 's/,//g' \
+                | awk '{print $1}')
+
+            L1D_LOADS=$(grep 'cpu_core/L1-dcache-loads' "$OUTPUT" \
                 | sed 's/,//g' \
                 | awk '{print $1}')
 
@@ -121,12 +126,12 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
             # Save to CSV
             # -------------------------------------------------
 
-            if [[ "$INSTRUCTIONS" =~ ^[0-9]+$ && "$L1D_MISSES" =~ ^[0-9]+$ ]]; then
-                echo "$MATRIXSIZE,$REGWIDTH,$INSTRUCTIONS,$L1D_MISSES,$SPEEDUP" >> "$CSV"
+            if [[ "$INSTRUCTIONS" =~ ^[0-9]+$ && "$L1D_MISSES" =~ ^[0-9]+$ && "$L1D_LOADS" =~ ^[0-9]+$ ]]; then
+                echo "$MATRIXSIZE,$REGWIDTH,$INSTRUCTIONS,$L1D_MISSES,$L1D_LOADS,$SPEEDUP" >> "$CSV"
                 rm -f "$OUTPUT"
                 break
             else
-                echo "Invalid instructions or L1-D misses value; rerunning matrix size $MATRIXSIZE, register width $REGWIDTH"
+                echo "Invalid instructions, L1-D misses, or L1-D loads value; rerunning matrix size $MATRIXSIZE, register width $REGWIDTH"
                 rm -f "$OUTPUT"
             fi
         done
@@ -135,6 +140,7 @@ for MATRIXSIZE in "${MATRIX_SIZES[@]}"; do
         echo "Captured:"
         echo "  Instructions : $INSTRUCTIONS"
         echo "  L1-D Misses  : $L1D_MISSES"
+        echo "  L1-D Loads   : $L1D_LOADS"
         echo "  Speedup      : $SPEEDUP"
 
     done
